--- app/helpers/oregon_digital/ability_helper.rb
@@ -1,21 +1,0 @@
-module OregonDigital
-  module AbilityHelper
-    def visibility_options(variant)
-      options = [
-        Hydra::AccessControls::AccessRight::VISIBILITY_TEXT_VALUE_PUBLIC,
-        'osu',
-        'uo',
-        Hydra::AccessControls::AccessRight::VISIBILITY_TEXT_VALUE_PRIVATE
-      ]
-      case variant
-      when :restrict
-        options.delete_at(0)
-        options.reverse!
-      when :loosen
-        options.delete_at(3)
-      end
-      options.map { |value| [visibility_text(value), value] }
-    end
-    
-  end
-end
