+++ app/presenters/hyrax/audio_presenter.rb
@@ -3,6 +3,5 @@
 module Hyrax
   # Display config for audio object
   class AudioPresenter < Hyrax::GenericPresenter
-    include OregonDigital::PresentsAttributes
   end
 end
