+++ app/presenters/hyrax/image_presenter.rb
@@ -3,7 +3,6 @@
 module Hyrax
   # Display config for image object
   class ImagePresenter < Hyrax::GenericPresenter
-    include OregonDigital::PresentsAttributes
     delegate(*Image.image_properties.map(&:to_sym), to: :solr_document)
   end
 end
