+++ app/presenters/hyrax/video_presenter.rb
@@ -3,7 +3,6 @@
 module Hyrax
   # Display config for video object
   class VideoPresenter < Hyrax::GenericPresenter
-    include OregonDigital::PresentsAttributes
     delegate(*Video.video_properties.map(&:to_sym), to: :solr_document)
   end
 end
