+++ web-workers.yml
@@ -1,0 +1,129 @@
+apiVersion: extensions/v1beta1
+kind: Deployment
+metadata:
+  annotations:
+    built_by: od2-infra/ansible/od2
+    deployment.kubernetes.io/revision: "8"
+    kubectl.kubernetes.io/last-applied-configuration: |
+      {"apiVersion":"extensions/v1beta1","kind":"Deployment","metadata":{"annotations":{"built_by":"od2-infra/ansible/od2","deployment.kubernetes.io/revision":"6"},"creationTimestamp":"2019-10-08T23:17:53Z","generation":6,"labels":{"k8s-app":"workers"},"name":"workers","namespace":"default","resourceVersion":"6493446","selfLink":"/apis/extensions/v1beta1/namespaces/default/deployments/workers","uid":"dac60117-ea21-11e9-b0bb-005056865b1e"},"spec":{"progressDeadlineSeconds":2147483647,"replicas":1,"revisionHistoryLimit":2147483647,"selector":{"matchLabels":{"k8s-app":"workers"}},"strategy":{"type":"Recreate"},"template":{"metadata":{"creationTimestamp":null,"labels":{"k8s-app":"workers"}},"spec":{"containers":[{"args":["bash","-c","./build/install_gems.sh \u0026\u0026 bundle exec rails assets:precompile \u0026\u0026 bundle exec sidekiq"],"env":[{"name":"ACTIVE_JOB_QUEUE_ADAPTER","value":"sidekiq"},{"name":"DATABASE_URL","value":"postgresql://postgres@db.default.svc.cluster.local/postgres"},{"name":"DEFAULT_URL_OPTION","value":"staging.oregondigital.org"},{"name":"DEPLOYED_VERSION","value":"WC3-0.0.1-5d504c8"},{"name":"FEDORA_URL","value":"http://fcrepo.default.svc.cluster.local:8080/fcrepo/rest"},{"name":"FITS_PATH","value":"/opt/fits-1.0.5/fits.sh"},{"name":"HONEYCOMB_DATASET","value":"od2-rails-staging"},{"name":"HONEYCOMB_SERVICE","value":"od2"},{"name":"HONEYCOMB_WRITEKEY","value":"d5529d94d4e989bb1fc1b1cfc559ebba"},{"name":"IIIF_SERVER_BASE_URL","value":"http://acaace0b9f1ca11e88634023959f013a-311393540.us-west-2.elb.amazonaws.com/iiif"},{"name":"RAILS_CACHE_STORE_URL","value":"memcache.default.svc.cluster.local"},{"name":"RAILS_ENV","value":"production"},{"name":"RAILS_LOG_TO_STDOUT","value":"true"},{"name":"RAILS_SERVE_STATIC_FILES","value":"true"},{"name":"REDIS_HOST","value":"redis.default.svc.cluster.local"},{"name":"REDIS_PORT","value":"6379"},{"name":"SECRET_KEY_BASE","value":"asdf"},{"name":"SOLR_URL","value":"http://solr.default.svc.cluster.local:8983/solr/staging"},{"name":"SMTP_ENABLED","value":"1"},{"name":"SMTP_HOST","value":"email-smtp.us-west-2.amazonaws.com"},{"name":"SMTP_PASSWD","value":"AveJNIObGL/gERpPb4vyOr9sL71mP1+jzZYuNc+ohxNC"},{"name":"SMTP_PORT","value":"587"},{"name":"SMTP_TLS","value":"true"},{"name":"SMTP_USER","value":"AKIAJHMMYVP5I6SW5JZA"},{"name":"SYSTEM_EMAIL_ADDRESS","value":"noreply@oregondigital.org"},{"name":"TRIPLESTORE_ADAPTER_TYPE","value":"blazegraph"},{"name":"TRIPLESTORE_ADAPTER_URL","value":"http://blazegraph.default.svc.cluster.local:8080/bigdata/namespace/kb/sparql"}],"image":"registry.library.oregonstate.edu/od2_web:20191018201601-WC3-0.0.1-2ea8c12","imagePullPolicy":"IfNotPresent","name":"workers","resources":{},"terminationMessagePath":"/dev/termination-log","terminationMessagePolicy":"File","volumeMounts":[{"mountPath":"/data/tmp/shared","name":"nfs-uploads"},{"mountPath":"/data/public/branding","name":"nfs-branding"}]}],"dnsPolicy":"ClusterFirst","restartPolicy":"Always","schedulerName":"default-scheduler","securityContext":{},"terminationGracePeriodSeconds":30,"volumes":[{"name":"nfs-uploads","persistentVolumeClaim":{"claimName":"nfs-uploads"}},{"name":"nfs-branding","persistentVolumeClaim":{"claimName":"nfs-branding"}}]}}},"status":{"availableReplicas":1,"conditions":[{"lastTransitionTime":"2019-10-18T22:35:33Z","lastUpdateTime":"2019-10-18T22:35:33Z","message":"Deployment has minimum availability.","reason":"MinimumReplicasAvailable","status":"True","type":"Available"}],"observedGeneration":6,"readyReplicas":1,"replicas":1,"updatedReplicas":1}}
+  creationTimestamp: 2019-10-08T23:17:53Z
+  generation: 8
+  labels:
+    k8s-app: workers
+  name: workers
+  namespace: default
+  resourceVersion: "7225460"
+  selfLink: /apis/extensions/v1beta1/namespaces/default/deployments/workers
+  uid: dac60117-ea21-11e9-b0bb-005056865b1e
+spec:
+  progressDeadlineSeconds: 2147483647
+  replicas: 1
+  revisionHistoryLimit: 2147483647
+  selector:
+    matchLabels:
+      k8s-app: workers
+  strategy:
+    type: Recreate
+  template:
+    metadata:
+      creationTimestamp: null
+      labels:
+        k8s-app: workers
+    spec:
+      containers:
+      - args:
+        - bash
+        - -c
+        - ./build/install_gems.sh && bundle exec rails assets:precompile && bundle
+          exec sidekiq
+        env:
+        - name: ACTIVE_JOB_QUEUE_ADAPTER
+          value: sidekiq
+        - name: DATABASE_URL
+          value: postgresql://postgres@db.default.svc.cluster.local/postgres
+        - name: DEFAULT_URL_OPTION
+          value: staging.oregondigital.org
+        - name: DEPLOYED_VERSION
+          value: WC3-0.0.1-5d504c8
+        - name: FEDORA_URL
+          value: http://fcrepo.default.svc.cluster.local:8080/fcrepo/rest
+        - name: FITS_PATH
+          value: /opt/fits-1.0.5/fits.sh
+        - name: HONEYCOMB_DATASET
+          value: od2-rails-staging
+        - name: HONEYCOMB_SERVICE
+          value: od2
+        - name: HONEYCOMB_WRITEKEY
+          value: d5529d94d4e989bb1fc1b1cfc559ebba
+        - name: IIIF_SERVER_BASE_URL
+          value: http://acaace0b9f1ca11e88634023959f013a-311393540.us-west-2.elb.amazonaws.com/iiif
+        - name: RAILS_CACHE_STORE_URL
+          value: memcache.default.svc.cluster.local
+        - name: RAILS_ENV
+          value: production
+        - name: RAILS_LOG_TO_STDOUT
+          value: "true"
+        - name: RAILS_SERVE_STATIC_FILES
+          value: "true"
+        - name: REDIS_HOST
+          value: redis.default.svc.cluster.local
+        - name: REDIS_PORT
+          value: "6379"
+        - name: SECRET_KEY_BASE
+          value: asdf
+        - name: SOLR_URL
+          value: http://solr.default.svc.cluster.local:8983/solr/staging
+        - name: SMTP_ENABLED
+          value: "1"
+        - name: SMTP_HOST
+          value: email-smtp.us-west-2.amazonaws.com
+        - name: SMTP_PASSWD
+          value: AveJNIObGL/gERpPb4vyOr9sL71mP1+jzZYuNc+ohxNC
+        - name: SMTP_PORT
+          value: "587"
+        - name: SMTP_TLS
+          value: "true"
+        - name: SMTP_USER
+          value: AKIAJHMMYVP5I6SW5JZA
+        - name: SYSTEM_EMAIL_ADDRESS
+          value: noreply@oregondigital.org
+        - name: TRIPLESTORE_ADAPTER_TYPE
+          value: blazegraph
+        - name: TRIPLESTORE_ADAPTER_URL
+          value: http://blazegraph.default.svc.cluster.local:8080/bigdata/namespace/kb/sparql
+        image:  registry.library.oregonstate.edu/od2_web:20191023165215-WC3-0.0.1-32aab30
+        imagePullPolicy: IfNotPresent
+        name: workers
+        resources: {}
+        terminationMessagePath: /dev/termination-log
+        terminationMessagePolicy: File
+        volumeMounts:
+        - mountPath: /data/tmp/shared
+          name: nfs-uploads
+        - mountPath: /data/public/branding
+          name: nfs-branding
+      dnsPolicy: ClusterFirst
+      restartPolicy: Always
+      schedulerName: default-scheduler
+      securityContext: {}
+      terminationGracePeriodSeconds: 30
+      volumes:
+      - name: nfs-uploads
+        persistentVolumeClaim:
+          claimName: nfs-uploads
+      - name: nfs-branding
+        persistentVolumeClaim:
+          claimName: nfs-branding
+status:
+  availableReplicas: 1
+  conditions:
+  - lastTransitionTime: 2019-10-21T18:08:25Z
+    lastUpdateTime: 2019-10-21T18:08:25Z
+    message: Deployment has minimum availability.
+    reason: MinimumReplicasAvailable
+    status: "True"
+    type: Available
+  observedGeneration: 8
+  readyReplicas: 1
+  replicas: 1
+  updatedReplicas: 1
